Add back usage of "reload eluna" command
With the upstream project rename to ALE, the "reload eluna" command was changed to "reload ale".

This is fine and will remain, but I'm also adding back the usage of "reload eluna" as a command as well.  Now you can use either "reload ale" or "reload eluna".  Both do the same thing.