Add optional strip_mask parameter to Loot::AddItem()
This is an optional parameter that can temporarily strip the bitmask of ItemTemplate flags when adding the item.

I found this to be useful in my aoe loot script around quest items that have a flag of 2048 due to how loot caching works.  It will allow the client display to update during this.